Add order creation
Create a new order. This is basically the same as creating an order through the REST API. The big difference is that this request returns an invoice ID for the order which can then be used to further manipulate the order.

The particular use case here is that I needed to create an order and apply the appropriate tax based on tax rules set up in Infusionsoft. The only way to do that that I could find was to create an order, then recalculate the tax on the invoice (using `InvoiceService.recalculateTax`).